Move to utils package
