+++ app/src/main/java/com/alphawallet/app/ui/SwapActivity.java
@@ -38,7 +38,7 @@ import com.alphawallet.app.ui.widget.entity.ProgressInfo;
 import com.alphawallet.app.util.BalanceUtils;
 import com.alphawallet.app.util.SwapUtils;
 import com.alphawallet.app.viewmodel.SwapViewModel;
-import com.alphawallet.app.viewmodel.LifiTokenUtils;
+import com.alphawallet.app.util.LifiTokenUtils;
 import com.alphawallet.app.web3.entity.Web3Transaction;
 import com.alphawallet.app.widget.AWalletAlertDialog;
 import com.alphawallet.app.widget.ActionSheetDialog;
